Click event via leap, select mesh, highlight mesh, rotate/resize or move with voice commands, add particles or animation
Added a new way of selecting objects/meshes: Click event, triggered if
finger (which is highlighted with a cursor), is at one position for a
certain amount of frames. Highlight that mesh (if any) and resize or
rotate the mesh, or say "verschieben" to move the mesh around, go back
to rotation/scaling or add animation or particles or let go of that mesh
and select another. Deselect only via voice command.